TopPage component includes nav and search
