+++ src/pages/Category/Category.js
@@ -3,22 +3,10 @@ import { useParams, useLocation } from 'react-router-dom';
 
 import { Box, Flex, Spinner } from '@chakra-ui/core';
 
-import { Search } from '../../components/Navigation';
 import { BookCategory } from '../../components/Book';
 
 import Pagination from './Pagination';
-import {
-    useBooksDispatch,
-    useDesignationBooks,
-    useDashboardHeightDispatch,
-} from '../../hooks/books';
-
-import {
-    useSearchSelector,
-    useSearchDispatch,
-    useSearchSelectedOptionDispatch,
-    useSearchPubFilterDispatch,
-} from '../../hooks/search';
+import { useBooksDispatch, useDesignationBooks } from '../../hooks/books';
 
 function getPage(query) {
     return Number(query.split('=')[1]);
@@ -27,19 +15,6 @@ function getPage(query) {
 export default function Category() {
     const { search } = useLocation();
     const { designation_id } = useParams();
-    const {
-        selectedOptions,
-        submitSelectedOption,
-        toggleObjects,
-    } = useSearchSelectedOptionDispatch();
-    const { updatePublisherFilter } = useSearchPubFilterDispatch();
-    const {
-        publisherInputTerm,
-        searchDetails,
-        selectedFilterOptions,
-        submitedFilterOption,
-    } = useSearchSelector();
-    const { changePageHeight, pageHeight } = useDashboardHeightDispatch();
     const page = getPage(search);
     const { loadDesignationPages } = useBooksDispatch();
 
@@ -53,41 +28,8 @@ export default function Category() {
         designation_id,
     });
 
-    const filterPublisherInput = e => {
-        e.preventDefault();
-        const searchedTerm = e.target.value;
-        updatePublisherFilter(searchedTerm);
-    };
-
-    const filteredOptions = (value, option) => {
-        selectedOptions(value, option);
-    };
-
-    const filterModalBackground = value => {
-        if (value) {
-            const HTMLheight = document.getElementsByTagName('html')[0]
-                .scrollHeight;
-            changePageHeight(`${HTMLheight}px`);
-        } else {
-            changePageHeight('0px');
-        }
-    };
-    const { setInput } = useSearchDispatch();
     return (
         <Box marginTop={['calc(75px + 2em)', 'calc(95px + 2em)']}>
-            <Search
-                onSearchInput={setInput}
-                searchDetails={searchDetails}
-                filteredOptions={filteredOptions}
-                filterPublisherInput={filterPublisherInput}
-                selectedFilterOptions={selectedFilterOptions}
-                filterModalBackground={filterModalBackground}
-                publisherInputTerm={publisherInputTerm}
-                pageHeight={pageHeight}
-                submitSelectedOption={submitSelectedOption}
-                submitedFilterOption={submitedFilterOption}
-                toggleObjects={toggleObjects}
-            />
             <Flex direction="column" alignItems="center">
                 {designationBooks.length <= 0 ? (
                     <Spinner marginTop="3rem" />
